pyproject.toml
[project]
name = "angela"
version = ""
description = ""
authors = [{ name = "rami3l", email = "rami3l@outlook.com" }]
dependencies = [
    "python-dotenv>=0.20.0",
    "aiogram>=2.20",
    "coloredlogs>=15.0.1",
    "overrides>=6.1.0",
    "aiohttp>=3.8.1",
    "certifi>=2021.10.8",
]
requires-python = ">=3.9"
license = { text = "MIT" }
[project.optional-dependencies]

[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"

[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "black>=22.3.0",
    "pytest>=7.1.1",
    "jurigged>=0.4.1",
    "mypy>=0.942",
    "pytest-asyncio>=0.18.3",
    "pydbg>=0.3.0",
]

[tool.pdm.scripts]
main = "python src/angela/main.py"
test = "pytest tests/"
fmt = "black ."
watch = "jurigged -v src/angela/main.py"
